♻️ 공통된 파이프라인 적용
